Major Refactor
No new functionality, but source code has been modified (See Issue #11)
