+++ .eslintignore
@@ -1,2 +1,2 @@
-build/
+app/
 public/
