+++ app/main.js
@@ -1,7 +1,6 @@
 const { app, BrowserWindow } = require('electron')
 
-function createWindow () {
-  // Create the browser window.
+app.on('ready', () => {
   const win = new BrowserWindow({
     width: 800,
     height: 600,
@@ -10,8 +9,5 @@ function createWindow () {
     }
   })
 
-  // and load the index.html of the app.
   win.loadURL('http://142.160.71.57:7777/')
-}
-
-app.on('ready', createWindow)
+})
