--- classes/utils.js
@@ -1,3 +1,0 @@
-module.exports = class Utils {
-  // Nothing in here at the moment.
-}
