+++ public/index.html
@@ -6,17 +6,17 @@
 	<link rel="stylesheet" href="css/style.css">
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
 
-	<script src="scripts/classes/utils.js"></script>
-	<script src="scripts/classes/entity.js"></script>
-	<script src="scripts/classes/livingentity.js"></script>
-	<script src="scripts/classes/player.js"></script>
-	<script src="scripts/classes/message.js"></script>
+	<script src="scripts/model/util.js"></script>
+	<script src="scripts/model/Entity.js"></script>
+	<script src="scripts/model/LivingEntity.js"></script>
+	<script src="scripts/model/Player.js"></script>
+	<script src="scripts/model/Message.js"></script>
+	<script src="scripts/model/Game.js"></script>
 
-	<script src="scripts/game.js"></script>
 	<script src="scripts/client.js"></script>
-	<script src="scripts/gui.js"></script>
+	<script src="scripts/ui.js"></script>
 
-	<script src="p5/p5.min.js"></script>
+	<script src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/0.9.0/p5.js"></script>
 	<script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/2.3.0/socket.io.dev.js"></script>
 	<!--Socket.io latest links can be found at https://cdnjs.com/libraries/socket.io-->
 </head>
