+++ public/scripts/model/Entity.js
