+++ public/scripts/model/Game.js
