+++ public/scripts/model/LivingEntity.js
