+++ public/scripts/model/Message.js
