--- start-app.cmd
@@ -1,3 +1,0 @@
-@ECHO OFF
-echo Running..
-electron --inspect=7777 app/main.js
